+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+Runtime changes:
+* Sanitize a literal "\`\`\`" line in snapshots ([#27](https://github.com/brandonchinn178/skeletest/issues/27))
+
 ## v0.3.7
 
 New features: